refactor: move checksum pacakge into encoding
