+++ encoding/checksum/scanner.go
