+++ encoding/checksum/scanner_test.go
