+++ encoding/checksum/types.go
